// web/src/lib/api/generated/models/poskasir-internal-orders-repository-order-type.ts
/* tslint:disable */
/* eslint-disable */
/**
 * POS Kasir API
 * POS Kasir API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */




export const POSKasirInternalOrdersRepositoryOrderType = {
    OrderTypeDineIn: 'dine_in',
    OrderTypeTakeaway: 'takeaway'
} as const;

export type POSKasirInternalOrdersRepositoryOrderType = typeof POSKasirInternalOrdersRepositoryOrderType[keyof typeof POSKasirInternalOrdersRepositoryOrderType];